+++ frontend/src/components/dialog/export/ExportDDL.tsx
@@ -284,6 +284,7 @@ const ExportDDL: React.FC<MenuDialogControl> = ({
                     { label: '删表语句', value: 'deleteTable' },
                     { label: '建表语句', value: 'createTable' },
                     { label: '建索引语句', value: 'createIndex' },
+                    { label: '建触发器语句', value: 'createTrigger' },
                     { label: '表注释语句', value: 'updateComment' },
                   ]}
                   onChange={(checkedValue) => {